// src/api/test_utils/enum.ts
export enum keylet_code {
  HOOK = 1,
  HOOK_STATE = 2,
  ACCOUNT = 3,
  AMENDMENTS = 4,
  CHILD = 5,
  SKIP = 6,
  FEES = 7,
  NEGATIVE_UNL = 8,
  LINE = 9,
  OFFER = 10,
  QUALITY = 11,
  EMITTED_DIR = 12,
  TICKET = 13,
  SIGNERS = 14,
  CHECK = 15,
  DEPOSIT_PREAUTH = 16,
  UNCHECKED = 17,
  OWNER_DIR = 18,
  PAGE = 19,
  ESCROW = 20,
  PAYCHAN = 21,
  EMITTED_TXN = 22,
  NFT_OFFER = 23,
  HOOK_DEFINITION = 24,
  HOOK_STATE_DIR = 25,
  LAST_KLTYPE_V0 = HOOK_DEFINITION,
  LAST_KLTYPE_V1 = HOOK_STATE_DIR,
}